Fix IdP Keycloak
After some investigation, the patch in the Makefile didn't get applied during the e2e test; therefore let's apply the patch to the repo and also save the patch as a file in case we decide to bump up the dependency in the future.

This commit also creates keycloak deployment instead of pod, so the pod will get recreated after a node restart. Previously, the node gets restarted after the remediation gets applied, and the keycloak pod will get terminated, therefore rendering authentication operator in a degreed state.